Support multiple player certificate keys and profile property keys
Allow authentication servers to specify an array of profile property
public keys and an array of player certificate public keys, similar to
Mojang's https://api.minecraftservices.com/publickeys route.

Adds two new fields, `playerCertificateKeys` and `profilePropertyKeys`,
to the authlib-injector metadata route at the root of the API. Both are
arrays of strings containing public keys in the same format as the
`signaturePublickey`. If `signaturePublickey` is specified, it will be
counted as both a player certificate key and a profile property key,
which is the status quo.

With this change, authentication server A can forward a texture payload
from authentication server B, and the client could accept a texture
signed by either authentication server.

This change also allows more flexibility for authentication server
operators.

Resolves https://github.com/yushijinhun/authlib-injector/issues/254.