Use empty publicKeySignature in dummy ProfileKeyFilter response
If we send an invalid signature here (`AA==`), then the vanilla server
will throw an error while handling the LoginHelloC2S packet, even if
`enforce-secure-profile` is `false` in `server.properties`. The client
will be disconnected with the message "Invalid signature for profile
public key. Try restarting your game."

It's better to send an empty string; the server will simply ignore the
missing publicKeySignature if `enforce-secure-profile` is false.

Resolves https://github.com/unmojang/drasl/issues/109.